% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockData.R
\name{fetch_stock_dataset}
\alias{fetch_stock_dataset}
\title{Fetch a timeseries of a list of stock_cd from stock_db}
\usage{
fetch_stock_dataset(stock_db, table_name, stock_cd_list)
}
\arguments{
\item{stock_db}{a stock database object to operate}

\item{table_name}{name of target table}

\item{stock_cd_list}{a list of stock_cd}
}
\value{
a A data frame on success, or NULL
}
\description{
Fetch a timeseries of a list of stock_cd from stock_db
}
