% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GtaData.R
\name{gta_db}
\alias{gta_db}
\title{Class creator of gta_db}
\usage{
gta_db(dsn = "GTA_SQLData")
}
\arguments{
\item{dsn}{odbc dsn string}
}
\value{
an object of gta_db class
}
\description{
Create a object for gta_db class
}
\examples{
gta <- gta_db()
}
