% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockData.R
\name{stock_db}
\alias{stock_db}
\title{Class creator of stock_db class}
\usage{
stock_db(stock_db_class, ...)
}
\arguments{
\item{stock_db_class}{class name of specific class of stock database,
e.g. gta_db}

\item{...}{addition params used by specific class of stock database(e.g. a dsn string
"GTA_SQLData")}
}
\value{
a object of class of stock db
}
\description{
Generic class creator of stock_db class
}
\examples{
stock_db <- stock_db(gta_db, "GTA_SQLData")

}
